+++ packages/canvas/src/canvas/BaseCanvasComponent.ts
@@ -17,8 +17,8 @@ import { ShapeNodeDefinition } from '@diagram-craft/canvas/shape/shapeNodeDefini
 import { EventKey } from '@diagram-craft/utils/event';
 import * as svg from '@diagram-craft/canvas/component/vdom-svg';
 import { Browser } from '@diagram-craft/canvas/browser';
-import * as html from './component/vdom-html';
-import { rawHTML, VNode } from './component/vdom';
+import * as html from '../component/vdom-html';
+import { rawHTML, VNode } from '../component/vdom';
 import styles from './canvas.css?inline';
 
 export type BaseCanvasProps = {
