+++ packages/canvas/src/canvas/EditableCanvasComponent.ts
@@ -1,15 +1,15 @@
-import { SelectionComponent } from './components/SelectionComponent';
-import { SelectionMarqueeComponent } from './components/SelectionMarqueeComponent';
-import { GridComponent } from './components/GridComponent';
-import { Actions, findAndExecuteAction } from './keyMap';
-import { DocumentBoundsComponent } from './components/DocumentBoundsComponent';
-import { DRAG_DROP_MANAGER, Modifiers } from './dragDropManager';
-import { BACKGROUND, Tool, ToolConstructor, ToolType } from './tool';
-import { DragLabelComponent } from './components/DragLabelComponent';
+import { SelectionComponent } from '../components/SelectionComponent';
+import { SelectionMarqueeComponent } from '../components/SelectionMarqueeComponent';
+import { GridComponent } from '../components/GridComponent';
+import { Actions, findAndExecuteAction } from '../keyMap';
+import { DocumentBoundsComponent } from '../components/DocumentBoundsComponent';
+import { DRAG_DROP_MANAGER, Modifiers } from '../dragDropManager';
+import { BACKGROUND, Tool, ToolConstructor, ToolType } from '../tool';
+import { DragLabelComponent } from '../components/DragLabelComponent';
 import { AnchorHandlesComponent } from '@diagram-craft/canvas/components/AnchorHandlesComponent';
-import { $cmp, createEffect, Observable } from './component/component';
-import * as svg from './component/vdom-svg';
-import * as html from './component/vdom-html';
+import { $cmp, createEffect, Observable } from '../component/component';
+import * as svg from '../component/vdom-svg';
+import * as html from '../component/vdom-html';
 import { Point } from '@diagram-craft/geometry/point';
 import { Box } from '@diagram-craft/geometry/box';
 import { Diagram } from '@diagram-craft/model/diagram';
@@ -17,10 +17,10 @@ import { ViewboxEvents } from '@diagram-craft/model/viewBox';
 import { DiagramElement, getTopMostNode, isNode } from '@diagram-craft/model/diagramElement';
 import { SelectionState, SelectionStateEvents } from '@diagram-craft/model/selectionState';
 import { EventHelper } from '@diagram-craft/utils/eventHelper';
-import { rawHTML } from './component/vdom';
+import { rawHTML } from '../component/vdom';
 import styles from './canvas.css?inline';
 import { PanTool } from '@diagram-craft/canvas-app/tools/panTool';
-import { Context } from './context';
+import { Context } from '../context';
 import { unique } from '@diagram-craft/utils/array';
 import { isResolvableToRegularLayer } from '@diagram-craft/model/diagramLayer';
 import { BaseCanvasComponent, BaseCanvasProps } from './BaseCanvasComponent';
