+++ packages/main/src/react-app/toolwindow/PickerToolWindow/ObjectPickerDrag.tsx
@@ -17,7 +17,7 @@ import { UnitOfWork } from '@diagram-craft/model/unitOfWork';
 import { DefaultStyles } from '@diagram-craft/model/diagramDefaults';
 import { deepClone } from '@diagram-craft/utils/object';
 import { clamp } from '@diagram-craft/utils/math';
-import { CanvasComponent } from '@diagram-craft/canvas/CanvasComponent';
+import { CanvasComponent } from '@diagram-craft/canvas/canvas/CanvasComponent';
 import { insert } from '@diagram-craft/canvas/component/vdom';
 import { registerStencilUse } from '@diagram-craft/canvas-app/recentStencils';
